Restructured pageable API
